Implemented position tracking
Can extract delta position from two images.

Doesn’t work as well if the camera is in the middle of adjusting
lighting (leaving a shadow into sunlight) and doesn’t account for
heading, but it’s a start.

Entach algorithms can be used to account for heading. Lighting changes
can be accounted for by tracking velocity. If instantaneous velocity is
vastly different, disregard the current data in exchange for derived
position (derive from velocity).