Started experimenting with serial on pyboard
One iteration caused the pyboard to crash. This is most likely because
of serial buffer overflow.
This is also probably what was causing the arduino to run slowly

Reverse engineered “Arduino-serial-messaging”
Will implement an easier to understand version suited for our project.
(Notes in Essential elements.rtf and My implementation plan.rtf)