Serial now completely works*
*until proven otherwise
Discovered that readSerialData() in SerialPacket.cpp has a condition
Serial.available()
If this becomes false (the packet hasn’t fully arrived), _inputChar
will be incomplete and will completed at a different iteration.
However, this counts as a serial event but not as a packet. When I was
pinging for data, the half arrived packet would count as a ping for
data  and not as a packet.
Now, there is a special ping character (‘x’).